+++ src/styles.ts
@@ -81,7 +81,6 @@ ${tag}:focus:after {
 }`;
 }
 
-// todo re-check why it's required inject per each element
 /** Style for small-scroll; vars --scroll, --scroll-hover to customize styling
  * @tutorial Troubleshooting
  * * cursor:pointer; doesn't work - this Chromium issue https://stackoverflow.com/questions/64402424/why-does-the-css-cursor-property-not-work-for-the-styled-scrollbar */