+++ test/jest/controls/control.pwd.test.js
@@ -129,6 +129,24 @@ describe("control.pwd", () => {
     expect(testEl.outerHTML).toMatchInlineSnapshot(
       `"<wup-pwd><label for="txt1"><span><input placeholder=" " type="password" id="txt1" autocomplete="new-password"><strong></strong><span class="wup-hidden">press Alt + V to show/hide password</span></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button><button eye="" aria-hidden="true" type="button" tabindex="-1"></button></label></wup-pwd>"`
     );
+
+    // cover case when input height changed on btnEye click
+    const inp = testEl.$refInput;
+    h.setupLayout(inp, { h: 44, y: 0, x: 0, w: 100 });
+    jest.spyOn(inp, "offsetHeight", "get").mockImplementation(() => {
+      if (inp.style.height) {
+        return +inp.style.height;
+      }
+      return inp.type === "password" ? 45 : 44;
+    });
+    testEl.$refBtnEye.click();
+    expect(inp.outerHTML).toMatchInlineSnapshot(
+      `"<input placeholder=" " type="text" id="txt1" autocomplete="new-password" style="height: 45px;">"`
+    );
+    testEl.$refBtnEye.click();
+    expect(inp.outerHTML).toMatchInlineSnapshot(
+      `"<input placeholder=" " type="password" id="txt1" autocomplete="new-password" style="height: 45px;">"`
+    );
   });
 
   test("storage not allowed", async () => {